requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0
pillow==9.5.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
tqdm==4.65.0
chardet==4.0.0